requirements.txt
# llama-index==0.11.16
sqlalchemy==2.0.37
anthropic==0.45.2
openai==1.61.1
httpx==0.28.1
python-dotenv==1.0.1
pydantic-settings>=2.0.0
langchain-openai==0.2.3
langchain-core==0.3.33
langchain-community==0.3.16
langchain-postgres==0.0.13
langgraph==0.2.69
streamlit==1.42.0
psycopg2-binary
tenacity>=8.2.0
